remove deprecated class: software.amazon.awssdk.core.util.StringUtils
The replacement class is: software.amazon.awssdk.utils.StringUtils
