+++ codegen/src/main/resources/macros/marshaller/json/ModelMarshallerMacro.ftl
@@ -15,7 +15,7 @@ import software.amazon.awssdk.core.exception.SdkClientException;
 import ${metadata.fullModelPackageName}.*;
 import software.amazon.awssdk.core.runtime.transform.Marshaller;
 import software.amazon.awssdk.utils.BinaryUtils;
-import software.amazon.awssdk.core.util.StringUtils;
+import software.amazon.awssdk.utils.StringUtils;
 import software.amazon.awssdk.core.util.IdempotentUtils;
 import software.amazon.awssdk.core.util.StringInputStream;
 import software.amazon.awssdk.core.protocol.json.*;
