+++ codegen/src/main/resources/templates/json/ModelJsonUnmarshaller.ftl
@@ -30,7 +30,7 @@ public class ${shape.shapeName}Unmarshaller implements Unmarshaller<${shape.shap
             if (context.getHeader("${memberModel.http.unmarshallLocationName}") != null) {
                 context.setCurrentHeader("${memberModel.http.unmarshallLocationName}");
                 <#if memberModel.variable.simpleType == "Instant">
-                    ${shape.variable.variableName}Builder.${memberModel.fluentSetterMethodName}(software.amazon.awssdk.core.util.DateUtils.parseRfc1123Date(context.readText()));
+                    ${shape.variable.variableName}Builder.${memberModel.fluentSetterMethodName}(software.amazon.awssdk.utils.DateUtils.parseRfc1123Date(context.readText()));
                 <#else>
                     ${shape.variable.variableName}Builder.${memberModel.fluentSetterMethodName}(<@MemberUnmarshallerDeclarationMacro.content memberModel />.unmarshall(context));
                 </#if>
