+++ codegen/src/main/resources/templates/json/RequestMarshaller.ftl
@@ -24,7 +24,7 @@ import ${metadata.fullModelPackageName}.*;
 import ${metadata.fullTransformPackageName}.*;
 import software.amazon.awssdk.core.runtime.transform.Marshaller;
 import software.amazon.awssdk.utils.BinaryUtils;
-import software.amazon.awssdk.core.util.StringUtils;
+import software.amazon.awssdk.utils.StringUtils;
 import software.amazon.awssdk.core.util.IdempotentUtils;
 import software.amazon.awssdk.core.util.StringInputStream;
 import software.amazon.awssdk.core.protocol.*;
