+++ core/src/main/java/software/amazon/awssdk/core/auth/HttpCredentialsProvider.java
@@ -26,8 +26,8 @@ import software.amazon.awssdk.core.exception.SdkClientException;
 import software.amazon.awssdk.core.internal.CredentialsEndpointProvider;
 import software.amazon.awssdk.core.internal.HttpCredentialsUtils;
 import software.amazon.awssdk.core.util.ComparableUtils;
-import software.amazon.awssdk.core.util.DateUtils;
 import software.amazon.awssdk.core.util.json.JacksonUtils;
+import software.amazon.awssdk.utils.DateUtils;
 import software.amazon.awssdk.utils.SdkAutoCloseable;
 import software.amazon.awssdk.utils.Validate;
 import software.amazon.awssdk.utils.cache.CachedSupplier;
