+++ core/src/main/java/software/amazon/awssdk/core/util/ClockSkewUtil.java
@@ -22,6 +22,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import software.amazon.awssdk.annotations.SdkInternalApi;
 import software.amazon.awssdk.core.http.HttpResponse;
+import software.amazon.awssdk.utils.DateUtils;
 
 @SdkInternalApi
 public final class ClockSkewUtil {
