+++ core/src/main/java/software/amazon/awssdk/core/util/Mimetypes.java
@@ -27,6 +27,7 @@ import java.util.StringTokenizer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import software.amazon.awssdk.utils.IoUtils;
+import software.amazon.awssdk.utils.StringUtils;
 
 /**
  * Utility class that maintains a listing of known Mimetypes, and determines the
