+++ services/sqs/src/it/java/software/amazon/awssdk/services/sqs/IntegrationTestBase.java
@@ -23,13 +23,13 @@ import java.util.Random;
 import java.util.UUID;
 import org.junit.Before;
 import software.amazon.awssdk.core.regions.Region;
-import software.amazon.awssdk.core.util.StringUtils;
 import software.amazon.awssdk.services.iam.IAMClient;
 import software.amazon.awssdk.services.iam.model.GetUserRequest;
 import software.amazon.awssdk.services.sqs.model.CreateQueueRequest;
 import software.amazon.awssdk.services.sqs.model.CreateQueueResponse;
 import software.amazon.awssdk.services.sqs.model.MessageAttributeValue;
 import software.amazon.awssdk.testutils.service.AwsIntegrationTestBase;
+import software.amazon.awssdk.utils.StringUtils;
 
 /**
  * Base class for SQS integration tests. Provides convenience methods for creating test data, and
@@ -159,7 +159,7 @@ public class IntegrationTestBase extends AwsIntegrationTestBase {
      */
     private String parseAccountIdFromArn(String arn) throws IllegalArgumentException {
         String[] arnComponents = arn.split(":");
-        if (arnComponents.length < 5 || StringUtils.isNullOrEmpty(arnComponents[4])) {
+        if (arnComponents.length < 5 || StringUtils.isEmpty(arnComponents[4])) {
             throw new IllegalArgumentException(String.format("%s is not a valid ARN", arn));
         }
         return arnComponents[4];
