+++ test/dynamodbmapper-v1/src/test/java/software/amazon/awssdk/services/dynamodb/datamodeling/StandardTypeConverters.java
@@ -39,8 +39,8 @@ import java.util.Set;
 import java.util.TimeZone;
 import java.util.regex.Pattern;
 import software.amazon.awssdk.annotations.SdkInternalApi;
-import software.amazon.awssdk.core.util.DateUtils;
 import software.amazon.awssdk.services.dynamodb.model.ScalarAttributeType;
+import software.amazon.awssdk.utils.DateUtils;
 
 /**
  * Type conversions.
