+++ test/dynamodbmapper-v1/src/test/java/software/amazon/awssdk/services/dynamodb/datamodeling/marshallers/DateSetToStringSetMarshaller.java
@@ -19,7 +19,7 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 import java.util.Set;
-import software.amazon.awssdk.core.util.DateUtils;
+import software.amazon.awssdk.utils.DateUtils;
 import software.amazon.awssdk.services.dynamodb.datamodeling.ArgumentMarshaller.StringSetAttributeMarshaller;
 import software.amazon.awssdk.services.dynamodb.model.AttributeValue;
