+++ test/dynamodbmapper-v1/src/test/java/software/amazon/awssdk/services/dynamodb/datamodeling/unmarshallers/CalendarSetUnmarshaller.java
@@ -20,8 +20,8 @@ import java.util.Calendar;
 import java.util.GregorianCalendar;
 import java.util.HashSet;
 import java.util.Set;
-import software.amazon.awssdk.core.util.DateUtils;
 import software.amazon.awssdk.services.dynamodb.model.AttributeValue;
+import software.amazon.awssdk.utils.DateUtils;
 
 /**
  * An unmarshaller that unmarshals sets of ISO-8601-formatted dates as sets of
