+++ test/dynamodbmapper-v1/src/test/java/software/amazon/awssdk/services/dynamodb/datamodeling/unmarshallers/DateSetUnmarshaller.java
@@ -18,7 +18,7 @@ package software.amazon.awssdk.services.dynamodb.datamodeling.unmarshallers;
 import java.util.Date;
 import java.util.HashSet;
 import java.util.Set;
-import software.amazon.awssdk.core.util.DateUtils;
+import software.amazon.awssdk.utils.DateUtils;
 import software.amazon.awssdk.services.dynamodb.model.AttributeValue;
 
 /**
