+++ test/protocol-tests-core/src/main/java/software/amazon/awssdk/protocol/asserts/marshalling/QueryParamsAssertion.java
@@ -29,7 +29,7 @@ import java.util.Map;
 import java.util.stream.Collectors;
 import org.apache.http.NameValuePair;
 import org.apache.http.client.utils.URLEncodedUtils;
-import software.amazon.awssdk.core.util.StringUtils;
+import software.amazon.awssdk.utils.StringUtils;
 
 /**
  * Asserts on the query parameters of the marshalled request.
@@ -89,7 +89,7 @@ public class QueryParamsAssertion extends MarshallingAssertion {
 
     private List<NameValuePair> parseNameValuePairsFromQuery(LoggedRequest actual) {
         final String queryParams = URI.create(actual.getUrl()).getQuery();
-        if (StringUtils.isNullOrEmpty(queryParams)) {
+        if (StringUtils.isEmpty(queryParams)) {
             return Collections.emptyList();
         }
         return URLEncodedUtils.parse(queryParams, StandardCharsets.UTF_8);
